Created separate test package for room related ops, renamed RoomLifeCycleManager related test file names to Room
