+++ chat-android/src/test/java/com/ably/chat/room/DetachTest.kt
@@ -1,0 +1,6 @@
+package com.ably.chat.room
+
+/**
+ * Spec: CHA-RL1
+ */
+class DetachTest
